requirements.txt
fastapi
uvicorn[standard]
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-google-genai
chromadb<1.0
langchain-chroma
pydantic
python-dotenv
